Use BnB selector only on first pass, knapsack selector on rest
Separate SelectCoinsMinConf to allow for either BnB selector or
knapsack selector. This allows CreateTransaction to run the BnB
selector on the first pass and the knapsack selector on the rest
to keep the original selection algorithm as a fallback.

Add a return parameter to selectcoins and selectoinsminconf to allow
the caller to know whether bnb was used for the coinselection or not.
This allows us to cleanly skip change output creation since BnB
would only be successful if there is no change.

Changed used_bnb and first_pass to be just first_pass renamed as use_bnb