Remove the CreateTransaction loop and fee checking things
We are now using effective value for coin selection, so there is no
need for the loop or any of the fee checking things that were done
within it. All fees ard handled by the effective value selection

Move vout filling to after coins are selected so that the transaction
fee is actually known to handle when users want to subtract the fee
from the outputs.