Remove use of isminefilters from the wallet
Instead of using isminefilters, use booleans to indicate whether
spendable or watch only is wanted. IsMine within the wallet is used
as a boolean. This means that IsMine is redefined as "any scriptPubKey
the wallet is tracking" so it encompasses both watch only and spendable.

For something to be spendable, it must be IsMine() and not watch only.

For something to be watch only, it must be in setWatchOnly, i.e.
HaveWatchOnly() returns true.